[indidetector] modified detector interface and added time deviation blob into the detector class

+++ libindi/examples/tutorial_eight/simple_detector_simulator.cpp
@@ -305,7 +305,7 @@ void SimpleDetector::grabFrame()
     PrimaryDetector.setSpectrumBufferSize(len);
  
    // Let's get a pointer to the frame buffer
-    double *spectrum = PrimaryDetector.getSpectrumBuffer();
+    uint8_t *spectrum = PrimaryDetector.getSpectrumBuffer();
 
     // Fill buffer with random pattern
     for (int i = 0; i < len; i++)
